requirements.txt
asgiref==3.9.1
Django==5.2.4
django-role-permissions==3.1.1
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
